tests moved and changeEmailAddress error message updated
