+++ api/tests/unit/users/test_unit_users_tasks.py
@@ -1,4 +1,9 @@
-from users.tasks import create_pipedrive_lead
+from django.core import mail
+
+from users.tasks import (
+    create_pipedrive_lead,
+    send_email_changed_notification_email,
+)
 
 
 def test_create_pipedrive_lead(mocker, admin_user):
@@ -11,3 +16,18 @@ def test_create_pipedrive_lead(mocker, admin_user):
 
     # Then
     mock_lead_tracker.create_lead.assert_called_once_with(admin_user)
+
+
+def test_send_email_changed_notification():
+    # When
+    send_email_changed_notification_email(
+        first_name="first_name",
+        from_email="fromtest@test.com",
+        original_email="test2@test.com",
+    )
+
+    # Then
+    assert len(mail.outbox) == 1
+    assert mail.outbox[0].subject == "Your Flagsmith email address has been changed"
+    assert mail.outbox[0].from_email == "fromtest@test.com"
+    assert mail.outbox[0].recipients() == ["test2@test.com"]
