+++ frontend/web/components/pages/AccountSettingsPage.js
@@ -141,47 +141,49 @@ class TheComponent extends Component {
                     />
                     <div className='col-md-8'>
                       <form className='mb-4' onSubmit={this.save}>
-                        <div className='md-8'>
-                          <InputGroup
-                            className='mt-2'
-                            title='Email Address'
-                            data-test='firstName'
-                            inputProps={{
-                              className: 'full-width',
-                              name: 'groupName',
-                              readOnly: true,
-                            }}
-                            value={email}
-                            onChange={(e) =>
-                              this.setState({
-                                first_name: Utils.safeParseEventValue(e),
-                              })
-                            }
-                            type='text'
-                            name='Email Address'
-                          />
-                          {Utils.getFlagsmithHasFeature('change_email') && (
-                            <div className='text-right mt-2'>
-                              <Button
-                                onClick={() =>
-                                  openModal(
-                                    'Change Email Address',
-                                    <ChangeEmailAddress
-                                      onComplete={() => {
-                                        closeModal()
-                                        AppActions.logout()
-                                      }}
-                                    />,
-                                  )
-                                }
-                                type='button'
-                                class='input-group-addon'
-                              >
-                                Change Email Address
-                              </Button>
-                            </div>
-                          )}
-                        </div>
+                        {Utils.getFlagsmithHasFeature('show_email_address') && (
+                          <div className='md-8'>
+                            <InputGroup
+                              className='mt-2'
+                              title='Email Address'
+                              data-test='firstName'
+                              inputProps={{
+                                className: 'full-width',
+                                name: 'groupName',
+                                readOnly: true,
+                              }}
+                              value={email}
+                              onChange={(e) =>
+                                this.setState({
+                                  first_name: Utils.safeParseEventValue(e),
+                                })
+                              }
+                              type='text'
+                              name='Email Address'
+                            />
+                            {Utils.getFlagsmithHasFeature('change_email') && (
+                              <div className='text-right mt-2'>
+                                <Button
+                                  onClick={() =>
+                                    openModal(
+                                      'Change Email Address',
+                                      <ChangeEmailAddress
+                                        onComplete={() => {
+                                          closeModal()
+                                          AppActions.logout()
+                                        }}
+                                      />,
+                                    )
+                                  }
+                                  type='button'
+                                  class='input-group-addon'
+                                >
+                                  Change Email Address
+                                </Button>
+                              </div>
+                            )}
+                          </div>
+                        )}
                         <InputGroup
                           className='mt-2'
                           title='First Name'
